Fixed some serialization errors for DrawnMapComponent objects.

+++ DrawnArrow.cs
@@ -194,25 +194,25 @@ namespace RealmStudio
             XAttribute? xAttr = arrowDoc.Root.Attribute("X");
             if (xAttr != null)
             {
-                X = int.Parse(xAttr.Value);
+                X = (int)float.Parse(xAttr.Value);
             }
 
             XAttribute? yAttr = arrowDoc.Root.Attribute("Y");
             if (yAttr != null)
             {
-                Y = int.Parse(yAttr.Value);
+                Y = (int)float.Parse(yAttr.Value);
             }
 
             XAttribute? wAttr = arrowDoc.Root.Attribute("Width");
             if (wAttr != null)
             {
-                Width = int.Parse(wAttr.Value);
+                Width = (int)float.Parse(wAttr.Value);
             }
 
             XAttribute? hAttr = arrowDoc.Root.Attribute("Height");
             if (hAttr != null)
             {
-                Height = int.Parse(hAttr.Value);
+                Height = (int)float.Parse(hAttr.Value);
             }
 
             XElement? topLeftElement = arrowDoc.Root.Element(ns + "TopLeft");
@@ -222,7 +222,7 @@ namespace RealmStudio
                 XAttribute? tlyAttr = topLeftElement.Attribute("Y");
                 if (tlxAttr != null && tlyAttr != null)
                 {
-                    _topLeft = new SKPoint(int.Parse(tlxAttr.Value), int.Parse(tlyAttr.Value));
+                    _topLeft = new SKPoint((int)float.Parse(tlxAttr.Value), (int)float.Parse(tlyAttr.Value));
                 }
             }
 
@@ -233,7 +233,7 @@ namespace RealmStudio
                 XAttribute? bryAttr = bottomRightElement.Attribute("Y");
                 if (brxAttr != null && bryAttr != null)
                 {
-                    _bottomRight = new SKPoint(int.Parse(brxAttr.Value), int.Parse(bryAttr.Value));
+                    _bottomRight = new SKPoint((int)float.Parse(brxAttr.Value), (int)float.Parse(bryAttr.Value));
                 }
             }
 
@@ -255,7 +255,7 @@ namespace RealmStudio
             XElement? rotationElement = arrowDoc.Root.Element(ns + "Rotation");
             if (rotationElement != null)
             {
-                _rotation = int.Parse(rotationElement.Value);
+                _rotation = (int)float.Parse(rotationElement.Value);
             }
 
             XElement? fillTypeElement = arrowDoc.Root.Element(ns + "FillType");
