+++ DrawnDiamond.cs
@@ -184,25 +184,25 @@ namespace RealmStudio
             XAttribute? xAttr = diamondDoc.Root.Attribute("X");
             if (xAttr != null)
             {
-                X = int.Parse(xAttr.Value);
+                X = (int)float.Parse(xAttr.Value);
             }
 
             XAttribute? yAttr = diamondDoc.Root.Attribute("Y");
             if (yAttr != null)
             {
-                Y = int.Parse(yAttr.Value);
+                Y = (int)float.Parse(yAttr.Value);
             }
 
             XAttribute? wAttr = diamondDoc.Root.Attribute("Width");
             if (wAttr != null)
             {
-                Width = int.Parse(wAttr.Value);
+                Width = (int)float.Parse(wAttr.Value);
             }
 
             XAttribute? hAttr = diamondDoc.Root.Attribute("Height");
             if (hAttr != null)
             {
-                Height = int.Parse(hAttr.Value);
+                Height = (int)float.Parse(hAttr.Value);
             }
 
             XElement? topLeftElement = diamondDoc.Root.Element(ns + "TopLeft");
@@ -212,7 +212,7 @@ namespace RealmStudio
                 XAttribute? tlyAttr = topLeftElement.Attribute("Y");
                 if (tlxAttr != null && tlyAttr != null)
                 {
-                    _topLeft = new SKPoint(int.Parse(tlxAttr.Value), int.Parse(tlyAttr.Value));
+                    _topLeft = new SKPoint((int)float.Parse(tlxAttr.Value), (int)float.Parse(tlyAttr.Value));
                 }
             }
 
@@ -223,7 +223,7 @@ namespace RealmStudio
                 XAttribute? bryAttr = bottomRightElement.Attribute("Y");
                 if (brxAttr != null && bryAttr != null)
                 {
-                    _bottomRight = new SKPoint(int.Parse(brxAttr.Value), int.Parse(bryAttr.Value));
+                    _bottomRight = new SKPoint((int)float.Parse(brxAttr.Value), (int)float.Parse(bryAttr.Value));
                 }
             }
 
@@ -239,13 +239,13 @@ namespace RealmStudio
             XElement? brushSizeElement = diamondDoc.Root.Element(ns + "BrushSize");
             if (brushSizeElement != null)
             {
-                _brushSize = int.Parse(brushSizeElement.Value);
+                _brushSize = (int)float.Parse(brushSizeElement.Value);
             }
 
             XElement? rotationElement = diamondDoc.Root.Element(ns + "Rotation");
             if (rotationElement != null)
             {
-                _rotation = int.Parse(rotationElement.Value);
+                _rotation = (int)float.Parse(rotationElement.Value);
             }
 
             XElement? fillTypeElement = diamondDoc.Root.Element(ns + "FillType");
