+++ DrawnRegularPolygon.cs
@@ -186,25 +186,25 @@ namespace RealmStudio
             XAttribute? xAttr = polyDoc.Root.Attribute("X");
             if (xAttr != null)
             {
-                X = int.Parse(xAttr.Value);
+                X = (int)float.Parse(xAttr.Value);
             }
 
             XAttribute? yAttr = polyDoc.Root.Attribute("Y");
             if (yAttr != null)
             {
-                Y = int.Parse(yAttr.Value);
+                Y = (int)float.Parse(yAttr.Value);
             }
 
             XAttribute? wAttr = polyDoc.Root.Attribute("Width");
             if (wAttr != null)
             {
-                Width = int.Parse(wAttr.Value);
+                Width = (int)float.Parse(wAttr.Value);
             }
 
             XAttribute? hAttr = polyDoc.Root.Attribute("Height");
             if (hAttr != null)
             {
-                Height = int.Parse(hAttr.Value);
+                Height = (int)float.Parse(hAttr.Value);
             }
 
             XElement? topLeftElement = polyDoc.Root.Element(ns + "TopLeft");
@@ -214,7 +214,7 @@ namespace RealmStudio
                 XAttribute? tlyAttr = topLeftElement.Attribute("Y");
                 if (tlxAttr != null && tlyAttr != null)
                 {
-                    _topLeft = new SKPoint(int.Parse(tlxAttr.Value), int.Parse(tlyAttr.Value));
+                    _topLeft = new SKPoint((int)float.Parse(tlxAttr.Value), (int)float.Parse(tlyAttr.Value));
                 }
             }
 
@@ -225,7 +225,7 @@ namespace RealmStudio
                 XAttribute? bryAttr = bottomRightElement.Attribute("Y");
                 if (brxAttr != null && bryAttr != null)
                 {
-                    _bottomRight = new SKPoint(int.Parse(brxAttr.Value), int.Parse(bryAttr.Value));
+                    _bottomRight = new SKPoint((int)float.Parse(brxAttr.Value), (int)float.Parse(bryAttr.Value));
                 }
             }
 
@@ -241,13 +241,13 @@ namespace RealmStudio
             XElement? brushSizeElement = polyDoc.Root.Element(ns + "BrushSize");
             if (brushSizeElement != null)
             {
-                _brushSize = int.Parse(brushSizeElement.Value);
+                _brushSize = (int)float.Parse(brushSizeElement.Value);
             }
 
             XElement? rotationElement = polyDoc.Root.Element(ns + "Rotation");
             if (rotationElement != null)
             {
-                _rotation = int.Parse(rotationElement.Value);
+                _rotation = (int)float.Parse(rotationElement.Value);
             }
 
             XElement? fillTypeElement = polyDoc.Root.Element(ns + "FillType");
@@ -259,7 +259,7 @@ namespace RealmStudio
             XElement? sidesElement = polyDoc.Root.Element(ns + "Sides");
             if (sidesElement != null)
             {
-                Sides = int.Parse(sidesElement.Value);
+                Sides = (int)float.Parse(sidesElement.Value);
             }
 
             XElement? fillBitmapElement = polyDoc.Root.Element(ns + "FillBitmap");
